float

+++ blend_matrix.py
@@ -1,13 +1,13 @@
 from typing import List, Set, Tuple
 
-def average(nums:List[int]) -> int:
+def average(nums:List[int]) -> float:
     """
     recives a non-empty list of integers,
     returns the average
     """
     return sum(nums)/len(nums)
 
-def get_neighbors_indexes(matrix:List[List[int]], index_row:int, index_column:int) -> Set[Tuple[int]]:
+def get_neighbors_indexes(matrix:List[List[float]], index_row:int, index_column:int) -> Set[Tuple[int]]:
     """
     recives a matrix with numerical values and two integers representing a position in the matrix
     returns a set of the element's neighbors' positions (represented as tuples)
@@ -25,14 +25,14 @@ def get_neighbors_indexes(matrix:List[List[int]], index_row:int, index_column:in
     return neighborhood
 
 
-def get_value_from_indexes(matrix:List[List[int]], index_row:int, index_column:int) -> int:
+def get_value_from_indexes(matrix:List[List[float]], index_row:int, index_column:int) -> float:
     """
     recives a matrix and two integers representing a position in the matrix
     returns the element of the matrix occupying that position
     """
     return matrix[index_row][index_column]
 
-def get_neighbors(matrix:List[List[int]], index_row:int, index_column:int) -> List[int]:
+def get_neighbors(matrix:List[List[float]], index_row:int, index_column:int) -> List[float]:
     """
     recives a matrix with numerical values and two integers representing a position in the matrix
     returns a list of the element's neighbors (the element is also included in the list)
@@ -41,7 +41,7 @@ def get_neighbors(matrix:List[List[int]], index_row:int, index_column:int) -> Li
     neighborhood = list(map(lambda x: get_value_from_indexes(matrix, x[0], x[1]), neighborhood_indexes))
     return neighborhood
 
-def process_element(matrix:List[List[int]], index_row:int, index_column:int) -> int:
+def process_element(matrix:List[List[float]], index_row:int, index_column:int) -> float:
     """
     recives a matrix with numerical values and two integers representing a position in the matrix
     returns the average between that element and its neighbors
@@ -49,7 +49,7 @@ def process_element(matrix:List[List[int]], index_row:int, index_column:int) -> 
     neighbors = get_neighbors(matrix, index_row, index_column)
     return average(neighbors)
 
-def process_matrix(matrix:List[List[int]]) -> List[List[int]]:
+def process_matrix(matrix:List[List[float]]) -> List[List[float]]:
     """
     recives a matrix with numerical values
     returns another matrix
